Fix SonarCloud findings (cycle 1)
Extract the schema compatibility differ into installation/_schema_compat.py
to bring publication.py under the file-length limit and decompose the
over-threshold differ functions into small, single-responsibility helpers
(cyclomatic <= 10, cognitive <= 15). Add docstrings, a _Defs type alias
replacing bare tuple[Mapping, Mapping] generics, object instead of Any for
schema-node params, a single-return boolean-schema normalizer, subset-based
type comparison, and a _DEFS_KEY constant to remove the duplicated "$defs"
literal.
