+++ shifter/installation/publication.py
@@ -37,6 +37,7 @@ from typing import Any
 import jsonschema
 from pydantic import ValidationError
 
+from ._schema_compat import find_incompatible_changes
 from .contract import (
     _BACKEND_NAME_RE,
     _PROFILE_RE,
@@ -201,271 +202,6 @@ def serialize_artifact(artifact: Mapping[str, Any]) -> str:
     return json.dumps(artifact, sort_keys=True, indent=2, ensure_ascii=False) + "\n"
 
 
-_REF_PREFIX = "#/$defs/"
-_MISSING = object()
-
-# JSON-Schema keyword pairs where tightening a bound rejects previously valid instances.
-# Lower bounds break when added or raised; upper bounds break when added or lowered.
-_LOWER_BOUND_KEYWORDS = ("minimum", "exclusiveMinimum", "minLength", "minItems", "minProperties")
-_UPPER_BOUND_KEYWORDS = ("maximum", "exclusiveMaximum", "maxLength", "maxItems", "maxProperties")
-
-# Validation keywords the differ reasons about explicitly. A change to any *other*
-# validation keyword is treated as potentially breaking (fail closed), so an unmodelled
-# Draft 2020-12 narrowing (for example ``dependentRequired`` or ``propertyNames``) cannot
-# slip through without a version bump.
-_HANDLED_KEYWORDS = frozenset(
-    {
-        "type",
-        "enum",
-        "const",
-        "properties",
-        "required",
-        "additionalProperties",
-        "items",
-        "prefixItems",
-        "anyOf",
-        "oneOf",
-        "allOf",
-        "not",
-        "pattern",
-        "multipleOf",
-        "uniqueItems",
-        "$ref",
-        *_LOWER_BOUND_KEYWORDS,
-        *_UPPER_BOUND_KEYWORDS,
-    }
-)
-# Annotation/metadata keywords that never affect validation — changing them is compatible.
-_ANNOTATION_KEYWORDS = frozenset(
-    {
-        "title",
-        "description",
-        "default",
-        "examples",
-        "readOnly",
-        "writeOnly",
-        "deprecated",
-        "$comment",
-        "$id",
-        "$schema",
-        "$defs",
-        "definitions",
-        "$anchor",
-    }
-)
-
-
-def _resolve_ref(node: Mapping[str, Any], defs: Mapping[str, Any]) -> Mapping[str, Any]:
-    """Resolve a local ``#/$defs/<name>`` reference to its definition (single level)."""
-    ref = node.get("$ref")
-    if isinstance(ref, str) and ref.startswith(_REF_PREFIX):
-        target = defs.get(ref.removeprefix(_REF_PREFIX))
-        if isinstance(target, Mapping):
-            return target
-    return node
-
-
-def _type_set(schema: Mapping[str, Any]) -> set[str] | None:
-    """The set of JSON types a schema allows, or ``None`` when unconstrained."""
-    value = schema.get("type")
-    if value is None:
-        return None
-    return {value} if isinstance(value, str) else set(value)
-
-
-def _diff_type_enum_const(old: Mapping[str, Any], new: Mapping[str, Any], path: str, changes: list[str]) -> None:
-    old_types, new_types = _type_set(old), _type_set(new)
-    if new_types is not None and old_types is None:
-        changes.append(f"{path}: type restricted to {sorted(new_types)} (was unconstrained)")
-    elif old_types is not None and new_types is not None and not old_types <= new_types:
-        changes.append(f"{path}: type(s) {sorted(old_types - new_types)} no longer accepted")
-    old_enum, new_enum = old.get("enum"), new.get("enum")
-    if old_enum is None and new_enum is not None:
-        changes.append(f"{path}: values restricted to a fixed enum")
-    elif old_enum is not None and new_enum is not None:
-        for value in old_enum:
-            if value not in new_enum:
-                changes.append(f"{path}: enum value {value!r} removed")
-    old_const, new_const = old.get("const", _MISSING), new.get("const", _MISSING)
-    if new_const is not _MISSING and new_const != old_const:
-        changes.append(f"{path}: const value added or changed")
-
-
-def _diff_constraints(old: Mapping[str, Any], new: Mapping[str, Any], path: str, changes: list[str]) -> None:
-    for keyword in _LOWER_BOUND_KEYWORDS:
-        o, n = old.get(keyword), new.get(keyword)
-        if n is not None and (o is None or n > o):
-            changes.append(f"{path}: {keyword} tightened")
-    for keyword in _UPPER_BOUND_KEYWORDS:
-        o, n = old.get(keyword), new.get(keyword)
-        if n is not None and (o is None or n < o):
-            changes.append(f"{path}: {keyword} tightened")
-    if new.get("pattern") is not None and new.get("pattern") != old.get("pattern"):
-        changes.append(f"{path}: pattern tightened")
-    if new.get("multipleOf") is not None and new.get("multipleOf") != old.get("multipleOf"):
-        changes.append(f"{path}: multipleOf tightened")
-    if new.get("uniqueItems") is True and old.get("uniqueItems") is not True:
-        changes.append(f"{path}: uniqueItems now required")
-    if new.get("not") is not None and new.get("not") != old.get("not"):
-        changes.append(f"{path}: 'not' constraint added or changed")
-
-
-def _diff_additional_properties(
-    old: Mapping[str, Any],
-    new: Mapping[str, Any],
-    defs: tuple[Mapping, Mapping],
-    path: str,
-    changes: list[str],
-    depth: int,
-) -> None:
-    old_ap = old.get("additionalProperties", True)
-    new_ap = new.get("additionalProperties", True)
-    old_permissive = old_ap is True or old_ap == {}
-    if new_ap is False and old_ap is not False:
-        changes.append(f"{path}: additionalProperties restricted to false")
-    elif isinstance(new_ap, Mapping) and new_ap != {}:
-        if old_permissive:
-            changes.append(f"{path}: additionalProperties restricted to a schema")
-        elif isinstance(old_ap, Mapping):
-            _diff_schema(old_ap, new_ap, defs, f"{path}.additionalProperties", changes, depth + 1)
-
-
-def _diff_object(
-    old: Mapping[str, Any],
-    new: Mapping[str, Any],
-    defs: tuple[Mapping, Mapping],
-    path: str,
-    changes: list[str],
-    depth: int,
-) -> None:
-    old_props = old.get("properties") or {}
-    new_props = new.get("properties") or {}
-    for name, old_prop in old_props.items():
-        if name not in new_props:
-            changes.append(f"{path}.{name}: property removed")
-        else:
-            _diff_schema(old_prop, new_props[name], defs, f"{path}.{name}", changes, depth + 1)
-    for name in sorted(set(new.get("required") or []) - set(old.get("required") or [])):
-        changes.append(f"{path}.{name}: newly required")
-    _diff_additional_properties(old, new, defs, path, changes, depth)
-
-
-def _diff_branches(
-    old: Mapping[str, Any],
-    new: Mapping[str, Any],
-    defs: tuple[Mapping, Mapping],
-    path: str,
-    changes: list[str],
-    depth: int,
-) -> None:
-    # anyOf/oneOf: fewer alternatives narrows the accepted set. allOf: MORE constraints narrows it.
-    # For all three, *adding* the keyword to a previously unconstrained schema also narrows.
-    for keyword in ("anyOf", "oneOf"):
-        old_b, new_b = old.get(keyword), new.get(keyword)
-        if isinstance(new_b, list) and not isinstance(old_b, list):
-            changes.append(f"{path}: {keyword} constraint added (was unconstrained)")
-        elif isinstance(old_b, list) and isinstance(new_b, list):
-            if len(new_b) < len(old_b):
-                changes.append(f"{path}: {keyword} narrowed ({len(old_b)} -> {len(new_b)} alternatives)")
-            for index, (o, n) in enumerate(zip(old_b, new_b, strict=False)):
-                _diff_schema(o, n, defs, f"{path}.{keyword}[{index}]", changes, depth + 1)
-    old_all, new_all = old.get("allOf"), new.get("allOf")
-    if isinstance(new_all, list) and not isinstance(old_all, list):
-        changes.append(f"{path}: allOf constraint added (was unconstrained)")
-    elif isinstance(old_all, list) and isinstance(new_all, list):
-        if len(new_all) > len(old_all):
-            changes.append(f"{path}: allOf narrowed ({len(old_all)} -> {len(new_all)} constraints)")
-        for index, (o, n) in enumerate(zip(old_all, new_all, strict=False)):
-            _diff_schema(o, n, defs, f"{path}.allOf[{index}]", changes, depth + 1)
-
-
-def _diff_items(
-    old: Mapping[str, Any],
-    new: Mapping[str, Any],
-    defs: tuple[Mapping, Mapping],
-    path: str,
-    changes: list[str],
-    depth: int,
-) -> None:
-    old_items, new_items = old.get("items"), new.get("items")
-    if isinstance(new_items, Mapping) and not isinstance(old_items, Mapping):
-        changes.append(f"{path}: items constraint added (was unconstrained)")
-    elif isinstance(old_items, Mapping) and isinstance(new_items, Mapping):
-        _diff_schema(old_items, new_items, defs, f"{path}.items", changes, depth + 1)
-    old_prefix, new_prefix = old.get("prefixItems"), new.get("prefixItems")
-    if isinstance(new_prefix, list) and not isinstance(old_prefix, list):
-        changes.append(f"{path}: prefixItems constraint added (was unconstrained)")
-    elif isinstance(old_prefix, list) and isinstance(new_prefix, list):
-        if len(new_prefix) > len(old_prefix):
-            changes.append(f"{path}: prefixItems added positional constraints")
-        for index, (o, n) in enumerate(zip(old_prefix, new_prefix, strict=False)):
-            _diff_schema(o, n, defs, f"{path}.prefixItems[{index}]", changes, depth + 1)
-
-
-def _diff_unknown_keywords(old: Mapping[str, Any], new: Mapping[str, Any], path: str, changes: list[str]) -> None:
-    """Fail closed: any changed validation keyword the differ does not model is a potential break."""
-    for keyword in sorted((set(old) | set(new)) - _HANDLED_KEYWORDS - _ANNOTATION_KEYWORDS):
-        if old.get(keyword, _MISSING) != new.get(keyword, _MISSING):
-            changes.append(f"{path}: unrecognized validation keyword {keyword!r} changed (treated as incompatible)")
-
-
-def _diff_boolean_schema(old: Any, new: Any, path: str, changes: list[str]) -> tuple[Any, Any] | None:
-    """Handle boolean subschemas (``true`` = accept all, ``false`` = accept none).
-
-    Returns the normalized ``(old, new)`` mapping pair to keep diffing, or ``None`` when the
-    comparison is already resolved (recorded a break, or a compatible widening).
-    """
-    if new is False and old is not False:
-        changes.append(f"{path}: now rejects all values")
-        return None
-    if new is True or old is False:
-        return None  # widening, or the old schema already rejected everything
-    old = {} if old is True else old
-    if not isinstance(old, Mapping) or not isinstance(new, Mapping):
-        return None
-    return old, new
-
-
-def _diff_schema(
-    old: Any, new: Any, defs: tuple[Mapping[str, Any], Mapping[str, Any]], path: str, changes: list[str], depth: int = 0
-) -> None:
-    """Recursively record where ``new`` rejects instances that ``old`` accepted."""
-    if depth > 64:
-        return
-    if isinstance(old, Mapping):
-        old = _resolve_ref(old, defs[0])
-    if isinstance(new, Mapping):
-        new = _resolve_ref(new, defs[1])
-    normalized = _diff_boolean_schema(old, new, path, changes)
-    if normalized is None:
-        return
-    old, new = normalized
-    _diff_type_enum_const(old, new, path, changes)
-    _diff_constraints(old, new, path, changes)
-    _diff_object(old, new, defs, path, changes, depth)
-    _diff_branches(old, new, defs, path, changes, depth)
-    _diff_items(old, new, defs, path, changes, depth)
-    _diff_unknown_keywords(old, new, path, changes)
-
-
-def find_incompatible_changes(old_schema: Mapping[str, Any], new_schema: Mapping[str, Any]) -> list[str]:
-    """Return the backward-incompatible changes from ``old_schema`` to ``new_schema``.
-
-    A change is backward-incompatible when ``new_schema`` rejects an instance that
-    ``old_schema`` accepted: a removed property, a removed enum value, a newly required
-    property, a narrowed type or union, a tightened numeric/string/array constraint, or
-    ``additionalProperties`` restricted to ``false``. The comparison recurses through
-    properties, ``$defs`` (resolving local ``$ref``), array ``items``, and
-    ``anyOf``/``oneOf``/``allOf`` branches, so nested schemas and validation narrowings are
-    caught, not just top-level fields. Additive changes are compatible and not reported.
-    Returns an empty list when ``new_schema`` is backward-compatible.
-    """
-    changes: list[str] = []
-    defs = (old_schema.get("$defs") or {}, new_schema.get("$defs") or {})
-    _diff_schema(old_schema, new_schema, defs, "root", changes)
-    return sorted(dict.fromkeys(changes))
-
-
 def _has_migration_note(migrations_path: Path, version: int) -> bool:
     """Whether ``MIGRATIONS.md`` carries a ``## Contract version <version>`` section."""
     if not migrations_path.exists():
@@ -487,6 +223,7 @@ def frozen_snapshots(directory: Path = PUBLISHED_CONTRACT_DIR) -> dict[int, Path
 
 
 def _drift_issues(generated_text: str, artifact_path: Path) -> list[ConfigIssue]:
+    """Report when the committed artifact is missing or out of date with the generated one."""
     if not artifact_path.exists():
         return [
             ConfigIssue(
